backend/requirements.txt
alabaster==0.7.12
asgiref==3.4.1
attrs==21.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
dateparser==1.1.0
Django==3.2.7
django-cors-headers==3.9.0
django-phone-field==1.8.1
djangorestframework==3.12.4
docutils==0.16
drf-spectacular==0.19.0
idna==3.2
imagesize==1.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==4.0.1
MarkupSafe==2.0.1
numpy==1.21.2
openapi-codec==1.3.2
packaging==21.0
pandas==1.3.3
psycopg2-binary==2.9.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2021.10.21
requests==2.26.0
simplejson==3.17.5
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
tzdata==2021.4
tzlocal==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
dj-rest-auth==2.1.11
django-allauth==0.47.0
